Clean code and optimize float conversion

+++ nebula/core/aggregation/fedavg.py
@@ -24,11 +24,12 @@ class FedAvg(Aggregator):
 
         last_model_params = models[-1][0]
         accum = {layer: torch.zeros_like(param, dtype=torch.float32) for layer, param in last_model_params.items()}
+        
         with torch.no_grad():
             for model_parameters, weight in models:
                 normalized_weight = weight / total_samples
                 for layer in accum:
-                    accum[layer].add_(model_parameters[layer].float(), alpha=normalized_weight)
+                    accum[layer].add_(model_parameters[layer].to(accum[layer].dtype), alpha=normalized_weight)
 
         # self.print_model_size(accum)
         return accum
